Everything is building except th JNI bridge
I had to make everything static libs as google protobuf does not do well in a dynamic library
cdm unit tests run and look good
pulse unit tests don't run via java testsuite, but seem ok (i ran cv only)
the engine does not run though... somthing to do with cardiac output not getting set, because the left heart to aorta path is zero?!

+++ test/cdm/cpp/UnitsTest.cpp
@@ -38,10 +38,10 @@ void CommonDataModelTest::ConvertList(std::vector<std::string> stringList)
   if (stringList.empty())
     return;
 
-  int size = stringList.size();
-  for (int i=0; i < size; i++)
+  size_t size = stringList.size();
+  for (size_t i=0; i < size; i++)
   {
-    for (int j=0; j < size; j++)
+    for (size_t j=0; j < size; j++)
     {
       if (i == j)
         continue;
